requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
djangorestframework
python-dotenv
openai